fixed client bug related to c-string

+++ trunk/main.cpp
@@ -31,39 +31,54 @@ void sig_child(int sig) {
 int main(int argc, char* argv[]) {
 	(void) signal(SIGINT,sig_child);
 
+	if (argc != 2) {
+		cerr << "Usage: " << argv[0] << " <port>" << endl;
+		return 1;
+	}
+
 	pid_t parent_ID = getpid();
 	pid_t pID = fork();
 	if (pID > 0) {
-		logger log(0,parent_ID);
-		log.write(5,"parent process beg");
-		log.write(0,to_str("forked child (pid = ") + to_str(pID) + to_str(")"));
+		logger* log = new logger(0,parent_ID);
+		log->write(5,"parent process beg");
+		log->write(0,to_str("forked child (pid = ") + to_str(pID) + to_str(")"));
 
-		string input;
-		while (true) {
-			if (child_exit)
-				exit(0);
+		{
+			log->write(0,"creating network client to server");
+			network::client cl("localhost",argv[1],log);
+			string line;
+			log->write(2,"beginning input block");
+			while (true) {
+				///\TODO:check if child process is dead by response of sending command to network socket
+				if (child_exit)
+					exit(0);
 
-			getline(cin,input);
-			if (cin.eof() || input.compare("exit") == 0)
-				break;
-			log.write(4,to_str("sending input to child process = ") + to_str(input));
-			///\TODO:check if child process is dead by response of sending command to network socket
+				cout << endl;
+				cout << "input: ";
+				getline(cin,line);
+
+				if (cin.eof() || line.compare("exit") == 0)
+					break;
+				log->write(4,to_str("sending input to child process = ") + to_str(line));
+				cl.send(line);
+			}
 		}
 
-		log.write(4,"request child process to terminate");
+		log->write(4,"request child process to terminate");
 		//send command to network socket
 		kill(pID,SIGINT);
 		//wait 3 sec, check for existence of child?  send kill command
 		waitpid(pID,0,0);
-		log.write(5,"parent process end");
+
+		log->write(5,"parent process end");
+		delete log;
 	} else if (pID == 0) {
 		logger log(0,getpid());
 		log.write(5,"child process beg");
 		dist::node* n = new dist::node(&log);
 
-
 		delete n;
-		//while (!child_exit) {;}
+		while (!child_exit) {;}
 		log.write(5,"child process end");
 	} else {
 		cerr << "Failed to fork" << endl;
